transform blottie vue to ts
