+++ src/index.ts
@@ -1,4 +1,4 @@
-import Blottie from './components/Blottie.vue'
+import Blottie from './components/Blottie'
 import { useBlottie } from './composables/blottie'
 import type { BlottieExpose, BlottiePlayer } from './typings/blottie'
